Separate tables for all tasks/job tasks
